:recycle: refactor: util 폴더명 utils로 변경
